ref(test): extract shared REPL-command test wiring into a trait
createReplTestIo() and prepareRunFactory() were copied across the REPL
integration tests. Hoist them into ReplCommandTestTrait so the new doc
test and its siblings share one definition.

Related to #2645
