+++ tests/php/Integration/Run/Command/Repl/ReplCommandTestTrait.php
@@ -1,0 +1,69 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PhelTest\Integration\Run\Command\Repl;
+
+use Gacela\Framework\Gacela;
+use Phel\Command\Application\TextExceptionPrinter;
+use Phel\Command\Domain\ErrorLogInterface;
+use Phel\Command\Domain\Exceptions\ExceptionArgsPrinter;
+use Phel\Command\Domain\Exceptions\Extractor\FilePositionExtractor;
+use Phel\Command\Infrastructure\SourceMapExtractor;
+use Phel\Run\Domain\Repl\ReplCommandIoInterface;
+use Phel\Run\RunFactory;
+use Phel\Shared\ColorStyle;
+use Phel\Shared\ColorStyleInterface;
+use Phel\Shared\Munge;
+use Phel\Shared\Printer\Printer;
+use Phel\Shared\Printer\PrinterInterface;
+use PHPUnit\Framework\TestCase;
+
+/**
+ * Shared REPL-command test wiring: a capturing {@see ReplTestIo} and a
+ * {@see RunFactory} override that routes REPL output through it.
+ *
+ * @psalm-require-extends TestCase
+ *
+ * @phpstan-require-extends TestCase
+ */
+trait ReplCommandTestTrait
+{
+    private function createReplTestIo(): ReplTestIo
+    {
+        $exceptionPrinter = new TextExceptionPrinter(
+            new ExceptionArgsPrinter(Printer::readable()),
+            ColorStyle::noStyles(),
+            new Munge(),
+            new FilePositionExtractor(new SourceMapExtractor()),
+            $this->createStub(ErrorLogInterface::class),
+        );
+
+        return new ReplTestIo($exceptionPrinter);
+    }
+
+    private function prepareRunFactory(ReplCommandIoInterface $io): void
+    {
+        Gacela::overrideExistingResolvedClass(
+            RunFactory::class,
+            new class($io) extends RunFactory {
+                public function __construct(private readonly ReplCommandIoInterface $io) {}
+
+                public function createColorStyle(): ColorStyleInterface
+                {
+                    return ColorStyle::noStyles();
+                }
+
+                public function createPrinter(): PrinterInterface
+                {
+                    return Printer::nonReadable();
+                }
+
+                public function createReplCommandIo(): ReplCommandIoInterface
+                {
+                    return $this->io;
+                }
+            },
+        );
+    }
+}
